[#1191] "Import Licenses" wizard should support import from archive
ImportLicenseDialog is extended with proper hooks
