+++ bundles/org.eclipse.passage.lic.jface/src/org/eclipse/passage/lic/internal/jface/dialogs/licensing/GoodIntention.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020, 2021 ArSysOp
+ * Copyright (c) 2020, 2024 ArSysOp
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -9,6 +9,7 @@
  *
  * Contributors:
  *     ArSysOp - initial API and implementation
+ *     ArSysOp - further support
  *******************************************************************************/
 package org.eclipse.passage.lic.internal.jface.dialogs.licensing;
 
@@ -27,7 +28,7 @@ public abstract class GoodIntention {
 	 */
 	public abstract boolean paveTheWay();
 
-	final static class Nope extends GoodIntention {
+	public final static class Nope extends GoodIntention {
 
 		@Override
 		public boolean paveTheWay() {
@@ -36,11 +37,11 @@ public abstract class GoodIntention {
 
 	}
 
-	final static class ImportLicense extends GoodIntention {
+	public final static class ImportLicense extends GoodIntention {
 
 		private final Supplier<Shell> shell;
 
-		ImportLicense(Supplier<Shell> shell) {
+		public ImportLicense(Supplier<Shell> shell) {
 			this.shell = shell;
 		}
 
@@ -50,11 +51,11 @@ public abstract class GoodIntention {
 		}
 	}
 
-	final static class RequestLicense extends GoodIntention {
+	public final static class RequestLicense extends GoodIntention {
 
 		private final Supplier<Shell> shell;
 
-		RequestLicense(Supplier<Shell> shell) {
+		public RequestLicense(Supplier<Shell> shell) {
 			this.shell = shell;
 		}
 
@@ -65,12 +66,12 @@ public abstract class GoodIntention {
 		}
 	}
 
-	final static class Diagnose extends GoodIntention {
+	public final static class Diagnose extends GoodIntention {
 
 		private final Supplier<Shell> shell;
 		private final Diagnostic diagnostic;
 
-		Diagnose(Supplier<Shell> shell, Diagnostic diagnostic) {
+		public Diagnose(Supplier<Shell> shell, Diagnostic diagnostic) {
 			this.shell = shell;
 			this.diagnostic = diagnostic;
 		}
@@ -82,12 +83,12 @@ public abstract class GoodIntention {
 		}
 	}
 
-	final static class ExposeLicenseAgreements extends GoodIntention {
+	public final static class ExposeLicenseAgreements extends GoodIntention {
 
 		private final Supplier<Shell> shell;
 		private final Collection<AgreementToAccept> agreements;
 
-		ExposeLicenseAgreements(Supplier<Shell> shell, Collection<AgreementToAccept> agreements) {
+		public ExposeLicenseAgreements(Supplier<Shell> shell, Collection<AgreementToAccept> agreements) {
 			this.shell = shell;
 			this.agreements = agreements;
 		}
